Drawing the board and pieces
I made a many files and clases thought I need them,
but just for making the board a need the last part of the file
src/board.py.
The other code was just for experimentation.

I need a little of documentation, refactoring and support
in a better way the fen notation.