Moving Pieces
Just the basic functionality of moving pieces.
Once the piece is pressed, the Sprite at that position is "disabled"
(set to None) and there's an active piece (the piece pressed)
then it can be moved due to the xy coordinates.
Once the piece is released the board Sprite at that position is
changed to the active piece (Sprite of course) and the
active piece is set to None, because there's no active piece to
drag and drop.

Also when the piece is released the alpha chanel is changed
reflecting the piece move.

TODO:
- Change the lightness if the pressed and released piece.
- Update the fen notation of the board.
- Maybe have a record of moves